feat: mv metrics to dedicated file
